+++ onnx/ops_test.go
@@ -251,3 +251,17 @@ func TestONNXDequantizeLinear(t *testing.T) {
 		[][]float32{{-3, 0, 300}, {-6, 90, 1200}},
 	}, -1)
 }
+
+func TestONNX_DynamicQuantizeLinear(t *testing.T) {
+	graphtest.RunTestGraphFn(t, "DequantizeLinear-scalar", func(g *Graph) (inputs, outputs []*Node) {
+		x := Const(g, [][]float32{{-3, -0, 3}, {-6, 9, 12}})
+		inputs = []*Node{x}
+		y, yScale, yZeroPoint := onnxDynamicQuantizeLinear(x)
+		outputs = []*Node{y, yScale, yZeroPoint}
+		return
+	}, []any{
+		[][]uint8{{43, 85, 127}, {0, 212, 255}},
+		float32(0.07058824),
+		uint8(85),
+	}, 1e-3)
+}