+++ docs/CHANGELOG.md
@@ -4,7 +4,7 @@
   Also added conversion when saving values back to the ONNX proto.
 * Updated dependencies to GoMLX.
 * Added `onnxtests.py` to help test/explore individual ONNX ops using ONNXRuntime. 
-* New ops: `DequantizeLinear`.
+* New ops: `DequantizeLinear`, `DynamicQuantizeLinear`.
 
 # v0.2.3 2025/05/31